+++ src/PipServices.Commons.NetStandard16/Config/FileConfigReader.cs
@@ -1,0 +1,19 @@
+﻿namespace PipServices.Commons.Config
+{
+    public abstract class FileConfigReader: IConfigReader, IConfigurable
+    {
+        public FileConfigReader(string path = null)
+        {
+            Path = path;
+        }
+
+        public string Path { get; set; }
+
+        public virtual void Configure(ConfigParams config)
+        {
+            Path = config.GetAsString("path");
+        }
+
+        public abstract ConfigParams ReadConfig(string correlationId);
+    }
+}